test : fold redundancy tests into agent detection module
Drops the standalone test_investigate_report_redundancy_e2e.py and
folds its 5 keeper tests as a TestReportRedundancy class in
test_agent_detection_e2e.py with a class-scoped second-EOT fixture.
Reuses the existing TestConfig load so the suite saves one scenario
reload and one EOT (~40s).

Adds Civic-Site Token artefact to setupTestConfig so Artefact_Searcher_Echo
reaches ARTEFACTSDIFF on a non-base location, letting the
artefact-outside-fold rule actually assert instead of skipping.

Drops the two HTML-element-presence tests that were subsumed by the
click-to-expand assertion.
